Make ExecutionEnv consistent across JVM/JS
