+++ common/js/src/main/scala/org/specs2/concurrent/ExecutionEnvPlatform.scala
@@ -5,19 +5,7 @@ import org.specs2.control.Logger
 import org.specs2.main.Arguments
 import scala.concurrent.*, duration.*
 
-/** Execution environment for javascript
-  */
-case class ExecutionEnv(executorServices: ExecutorServices, timeFactor: Int):
-
-  def shutdown(): Unit = ()
-
-  given executionContext: ExecutionContext = executorServices.executionContext
-  given ec: ExecutionContext = executionContext
-
-  def schedule(action: =>Unit, duration: FiniteDuration): Unit =
-    executorServices.schedule(action, duration)
-
-object ExecutionEnv:
+private[concurrent] trait ExecutionEnvCompanionPlatform:
 
   def create(arguments: Arguments, systemLogger: Logger, tag: Option[String] = None): ExecutionEnv =
     createSpecs2(arguments, systemLogger, tag)
